feat: add precision-dependent lag to ADC measurements.
The on-board ADC can be set to different resolutions, and at higher resolutions, the sampling rate is much lower.
Thus, we set the sampling rate of the ADC to make sure that enough time elapses in reading to ensure the full precision.

Merge @awickert branch of this into mine, build on top

fix: repeated lines